requirements.txt
asgiref==3.8.1
certifi==2025.6.15
charset-normalizer==3.4.2
dj-database-url==3.0.1
Django==5.2.3
django-cors-headers==4.7.0
django-sendgrid-v5==1.3.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
ecdsa==0.19.1
et_xmlfile==2.0.0
gunicorn==23.0.0
idna==3.10
MarkupSafe==3.0.2
openpyxl==3.1.5
packaging==25.0
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-http-client==3.3.7
PyYAML==6.0.2
requests==2.32.4
sendgrid==3.6.5
sendgrid-django==4.2.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
ua-parser==1.0.1
ua-parser-builtins==0.18.0.post1
urllib3==2.5.0
user-agents==2.2.0
uuid==1.30
Werkzeug==3.1.3
whitenoise==6.9.0
